Create and Extend AuthenticateTask Abstract Class
