+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/backgroundTask/AuthenticateTask.java
@@ -1,0 +1,48 @@
+package edu.byu.cs.tweeter.client.model.service.backgroundTask;
+
+import android.os.Bundle;
+import android.os.Handler;
+
+import edu.byu.cs.tweeter.model.domain.AuthToken;
+import edu.byu.cs.tweeter.model.domain.User;
+import edu.byu.cs.tweeter.util.Pair;
+
+public abstract class AuthenticateTask extends BackgroundTask {
+
+    public static final String USER_KEY = "user";
+    public static final String AUTH_TOKEN_KEY = "auth-token";
+
+    /**
+     * The user's username (or "alias" or "handle"). E.g., "@susan".
+     */
+    private String username;
+    /**
+     * The user's password.
+     */
+    private String password;
+
+    private User authenticatedUser;
+    private AuthToken authToken;
+
+    public AuthenticateTask(Handler messageHandler, String username, String password) {
+        super(messageHandler);
+        this.username = username;
+        this.password = password;
+    }
+
+    @Override
+    protected void executeTask() {
+        Pair<User, AuthToken> authenticationResult = doAuthentication();
+
+        authenticatedUser = authenticationResult.getFirst();
+        authToken = authenticationResult.getSecond();
+    }
+
+    protected abstract Pair<User, AuthToken> doAuthentication();
+
+    @Override
+    protected void loadSuccessBundle(Bundle msgBundle) {
+        msgBundle.putSerializable(USER_KEY, authenticatedUser);
+        msgBundle.putSerializable(AUTH_TOKEN_KEY, authToken);
+    }
+}
