+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/backgroundTask/LoginTask.java
@@ -10,46 +10,16 @@ import edu.byu.cs.tweeter.util.Pair;
 /**
  * Background task that logs in a user (i.e., starts a session).
  */
-public class LoginTask extends BackgroundTask {
-
+public class LoginTask extends AuthenticateTask {
     private static final String LOG_TAG = "LoginTask";
 
-    public static final String USER_KEY = "user";
-    public static final String AUTH_TOKEN_KEY = "auth-token";
-
-    /**
-     * The user's username (or "alias" or "handle"). E.g., "@susan".
-     */
-    private String username;
-    /**
-     * The user's password.
-     */
-    private String password;
-
-    private User loggedInUser;
-    private AuthToken authToken;
-
     public LoginTask(String username, String password, Handler messageHandler) {
-        super(messageHandler);
-        this.username = username;
-        this.password = password;
+        super(messageHandler, username, password);
     }
 
     @Override
-    protected void executeTask() {
-        Pair<User, AuthToken> loginResult = doLogin();
-
-        loggedInUser = loginResult.getFirst();
-        authToken = loginResult.getSecond();
-    }
-
-    @Override
-    protected void loadSuccessBundle(Bundle msgBundle) {
-        msgBundle.putSerializable(USER_KEY, loggedInUser);
-        msgBundle.putSerializable(AUTH_TOKEN_KEY, authToken);
-    }
-
-    private Pair<User, AuthToken> doLogin() {
+    protected Pair<User, AuthToken> doAuthentication() {
+        // TODO: This will differ from RegisterTask's implementation when we are no longer using fake data
         User loggedInUser = getFakeData().getFirstUser();
         AuthToken authToken = getFakeData().getAuthToken();
         return new Pair<>(loggedInUser, authToken);
